Extracts attribute functionality to a new Attribute class in the DOM Module.
