+++ inc/Engine/CriticalPath/DOM.php
@@ -92,48 +92,6 @@ class DOM {
 	}
 
 	/**
-	 * Converts an array into a string.
-	 *
-	 * @since 3.6.2
-	 *
-	 * @param array  $array    The array to convert.
-	 * @param string $glue     Glue between the string parts.
-	 * @param string $operator Operator between the key and value when flatten the array.
-	 *
-	 * @return string converted string.
-	 */
-	protected function array_to_string( array $array, $glue, $operator ) {
-		return implode( $glue, $this->flatten_array( $array, $operator ) );
-	}
-
-	/**
-	 * Flattens an array from key => value to string elements.
-	 *
-	 * For index key, the value is stored as the element.
-	 * For keys, the key is combined with the value with the operator as the separator.
-	 *
-	 * @since 3.6.2
-	 *
-	 * @param array  $array    The array to flatten.
-	 * @param string $operator The separator between the key and value.
-	 *
-	 * @return array
-	 */
-	protected function flatten_array( array $array, $operator ) {
-		$flat = [];
-
-		foreach ( $array as $key => $value ) {
-			if ( is_integer( $key ) ) {
-				$flat[] = $value;
-			} else {
-				$flat[] = "{$key}{$operator}{$value}";
-			}
-		}
-
-		return $flat;
-	}
-
-	/**
 	 * Sets the <noscript>.
 	 *
 	 * @since 3.6.2
@@ -152,100 +110,5 @@ class DOM {
 		if ( $need_to_create ) {
 			$this->dom->get_body()->appendChild( $this->noscript );
 		}
-	}
-
-	/**
-	 * Prepares the given attribute value for embedding into the attribute.
-	 *
-	 * If it's a string, it's wrapped in quotations.
-	 *
-	 * @since 3.6.2
-	 *
-	 * @param mixed $value Value to wrap.
-	 *
-	 * @return mixed
-	 */
-	protected function prepare_for_value_embed( $value ) {
-		if ( $this->is_null( $value ) ) {
-			return 'null';
-		}
-
-		if ( empty( $value ) ) {
-			return '';
-		}
-
-		if ( ! is_string( $value ) ) {
-			return $value;
-		}
-
-		if ( "'" === $value[0] ) {
-			return $value;
-		}
-
-		if ( '"' === $value[0] ) {
-			return $this->replace_double_quotes( $value );
-		}
-
-		return "'{$value}'";
-	}
-
-	/**
-	 * Checks if the given value is set to NULL, 'null', "null", or null.
-	 *
-	 * @since 3.6.2
-	 *
-	 * @param string $value Value to check.
-	 *
-	 * @return bool
-	 */
-	public static function is_null( $value ) {
-		return (
-			is_null( $value )
-			||
-			'null' === $value
-			||
-			"null" === $value // phpcs:ignore Squiz.Strings.DoubleQuoteUsage.NotRequired -- This is intentional.
-		);
-	}
-
-	/**
-	 * Replaces double "" quotes with single quotes ''.
-	 *
-	 * @since 1.0.0
-	 *
-	 * @param string $value Value to replace double quotes.
-	 *
-	 * @return string value with single quotes around it instead of double.
-	 */
-	public static function replace_double_quotes( $value ) {
-		return str_replace( '"', "'", $value );
-	}
-
-	/**
-	 * Checks if the string contains the given needle.
-	 *
-	 * @since 3.6.2
-	 *
-	 * @param string $search_string Search string.
-	 * @param string $needle        Needle to find.
-	 *
-	 * @return bool
-	 */
-	public static function string_contains( $search_string, $needle ) {
-		return ( false !== strpos( $search_string, $needle ) );
-	}
-
-	/**
-	 * Checks if the search string starts with the given needle.
-	 *
-	 * @since 3.6.2
-	 *
-	 * @param string $search_string Search string.
-	 * @param string $needle        Needle to find.
-	 *
-	 * @return bool
-	 */
-	protected function string_starts_with( $search_string, $needle ) {
-		return ( substr( $search_string, 0, strlen( $needle ) ) === $needle );
 	}
 }
